# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AppVault - App Discovery Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AppAppType = (typeof AppAppType)[keyof typeof AppAppType];

export const AppAppType = {
  app: "app",
  game: "game",
} as const;

export type AppPlatform = (typeof AppPlatform)[keyof typeof AppPlatform];

export const AppPlatform = {
  ios: "ios",
  android: "android",
  both: "both",
} as const;

export interface App {
  id: number;
  name: string;
  developer: string;
  description: string;
  shortDescription: string;
  iconUrl: string;
  screenshotUrls: string[];
  categorySlug: string;
  categoryName: string;
  appType: AppAppType;
  rating: number;
  reviewCount: number;
  downloadCount: number;
  price: number;
  isFree: boolean;
  platform: AppPlatform;
  featured: boolean;
  trending: boolean;
  appStoreUrl?: string | null;
  playStoreUrl?: string | null;
  tags: string[];
  createdAt: string;
}

export type AppDetail = App & {
  fullDescription?: string;
  version?: string;
  size?: string;
  requirements?: string;
  website?: string | null;
  privacyPolicyUrl?: string | null;
  releaseNotes?: string | null;
  updatedAt?: string;
};

export type CategoryType = (typeof CategoryType)[keyof typeof CategoryType];

export const CategoryType = {
  app: "app",
  game: "game",
  all: "all",
} as const;

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  iconName: string;
  appCount: number;
  color: string;
  type: CategoryType;
}

export interface StatsSummary {
  totalApps: number;
  totalGames: number;
  totalCategories: number;
  totalDownloads: number;
  featuredCount: number;
}

export interface ErrorResponse {
  error: string;
}

export type ListAppsParams = {
  category?: string;
  search?: string;
  featured?: boolean;
  trending?: boolean;
  platform?: ListAppsPlatform;
  appType?: ListAppsAppType;
  limit?: number;
  offset?: number;
};

export type ListAppsPlatform =
  (typeof ListAppsPlatform)[keyof typeof ListAppsPlatform];

export const ListAppsPlatform = {
  ios: "ios",
  android: "android",
  both: "both",
} as const;

export type ListAppsAppType =
  (typeof ListAppsAppType)[keyof typeof ListAppsAppType];

export const ListAppsAppType = {
  app: "app",
  game: "game",
} as const;

export type GetFeaturedAppsParams = {
  appType?: GetFeaturedAppsAppType;
};

export type GetFeaturedAppsAppType =
  (typeof GetFeaturedAppsAppType)[keyof typeof GetFeaturedAppsAppType];

export const GetFeaturedAppsAppType = {
  app: "app",
  game: "game",
} as const;

export type GetTrendingAppsParams = {
  appType?: GetTrendingAppsAppType;
  limit?: number;
};

export type GetTrendingAppsAppType =
  (typeof GetTrendingAppsAppType)[keyof typeof GetTrendingAppsAppType];

export const GetTrendingAppsAppType = {
  app: "app",
  game: "game",
} as const;

export type GetNewAppsParams = {
  appType?: GetNewAppsAppType;
  limit?: number;
};

export type GetNewAppsAppType =
  (typeof GetNewAppsAppType)[keyof typeof GetNewAppsAppType];

export const GetNewAppsAppType = {
  app: "app",
  game: "game",
} as const;

export type GetPopularGamesParams = {
  limit?: number;
};

export type GetRelatedAppsParams = {
  limit?: number;
};

export type ListCategoriesParams = {
  type?: ListCategoriesType;
};

export type ListCategoriesType =
  (typeof ListCategoriesType)[keyof typeof ListCategoriesType];

export const ListCategoriesType = {
  app: "app",
  game: "game",
  all: "all",
} as const;

export type GetAppsByCategoryParams = {
  limit?: number;
  offset?: number;
};