# lib/api-zod/src/generated/types/app.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AppVault - App Discovery Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AppAppType } from "./appAppType";
import type { AppPlatform } from "./appPlatform";

export interface App {
  id: number;
  name: string;
  developer: string;
  description: string;
  shortDescription: string;
  iconUrl: string;
  screenshotUrls: string[];
  categorySlug: string;
  categoryName: string;
  appType: AppAppType;
  rating: number;
  reviewCount: number;
  downloadCount: number;
  price: number;
  isFree: boolean;
  platform: AppPlatform;
  featured: boolean;
  trending: boolean;
  appStoreUrl?: string | null;
  playStoreUrl?: string | null;
  tags: string[];
  createdAt: Date;
}