# lib/api-zod/src/generated/types/appDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AppVault - App Discovery Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { App } from "./app";

export type AppDetail = App & {
  fullDescription?: string;
  version?: string;
  size?: string;
  requirements?: string;
  website?: string | null;
  privacyPolicyUrl?: string | null;
  releaseNotes?: string | null;
  updatedAt?: Date;
};