# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AppVault - App Discovery Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryType } from "./categoryType";

export interface Category {
  id: number;
  name: string;
  slug: string;
  description: string;
  iconName: string;
  appCount: number;
  color: string;
  type: CategoryType;
}