# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AppVault - App Discovery Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalApps: number;
  totalGames: number;
  totalCategories: number;
  totalDownloads: number;
  featuredCount: number;
}